feat: 카드 구매자 상세 페이지 밑에 교환정보 조금 추가
