+++ src/components/HompageComponents/BaseCardsSection.jsx
@@ -7,7 +7,7 @@ import HomeFallbackCount from "../skeletons/HomeFallbackCount";
 import NoResultMessage from "../ui/NoResultsMessage";
 
 export default async function BaseCardsSection({ grade, genre, sale }) {
-  const API_BASE_URL = "http://localhost:3002";
+  const API_BASE_URL = "https://six-favoritephoto-4team-be.onrender.com";
   
   const res = await fetch(`${API_BASE_URL}/api/store`, {
     next: { revalidate: 60 },
@@ -16,7 +16,6 @@ export default async function BaseCardsSection({ grade, genre, sale }) {
     throw new Error("카드 데이터를 불러오는데 실패했습니다.");
   }
   const data = await res.json();
-  console.log(data)
   // 다중 선택 가능하도록 파싱 함수 (빈 배열 반환)
   function parseFilterValue(value) {
     if (!value) return [];
