+++ src/components/ui/BaseCardList.jsx
@@ -3,7 +3,6 @@ import BaseCard from './BaseCard';
 
 export default async function BaseCardList({ cards }) {
   await delay(2000)
-  console.log("💥 cards in BaseCardList:", cards); 
   return (
     <>
       {cards.map((card, index) => (
