+++ src/lib/api/fetch-client.js
@@ -5,7 +5,7 @@
 //  import { useAuth } from "@/providers/AuthProvider"; 이후에
 //  const { user, isLoading } = useAuth(); 하고 if (user) 면 로그인 상태 if (!user)면 로그아웃 상태
 
-const API_BASE_URL = "http://localhost:3002";
+const API_BASE_URL = "https://six-favoritephoto-4team-be.onrender.com";
 
 export const authUtils = {
   setAccessToken: (accessToken) => {
